TC-60: ServiceRequestCreateRequest Abstraction
