+++ src/main/java/edu/wpi/cs3733/D22/teamC/App.java
@@ -16,9 +16,9 @@ public class App extends Application {
     public static App instance;
 
     // Constants
-    private final String BASE_VIEW_PATH = "view/general/base-view.fxml";
-    private final String MENU_BAR_COMPONENT_PATH = "component/menu-bar.fxml";
-    private final String MEDICAL_EQUIPMENT = "view/service_request/medical-equipment-view.fxml";
+    public static final String BASE_VIEW_PATH = "view/general/base-view.fxml";
+    private static final String MENU_BAR_COMPONENT_PATH = "component/menu-bar.fxml";
+    private static final String MEDICAL_EQUIPMENT = "view/service_request/medical-equipment-view.fxml";
 
     // Variables
     private Stage stage;
