+++ src/main/resources/edu/wpi/cs3733/D22/teamC/view/service_request/medical-equipment-view.fxml
@@ -11,7 +11,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 
-<ScrollPane prefWidth="602.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.wpi.cs3733.D22.teamC.controller.service_request.MedicalEquipmentController">
+<ScrollPane prefWidth="602.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.wpi.cs3733.D22.teamC.controller.service_request.MedicalEquipmentSRCreateController">
    <content>
       <AnchorPane prefHeight="667.0" prefWidth="600.0">
          <children>
